moved progressMessengers definitions
