+++ src/Oceanostics.jl
@@ -84,7 +84,7 @@ using Oceananigans.TurbulenceClosures: νᶜᶜᶜ, calc_nonlinear_κᶜᶜᶜ
 include("TKEBudgetTerms.jl")
 include("TracerVarianceBudgetTerms.jl")
 include("FlowDiagnostics.jl")
-include("ProgressMessengers.jl")
+include("ProgressMessengers/ProgressMessengers.jl")
 include("progress_messengers.jl")
 
 using .TKEBudgetTerms, .TracerVarianceBudgetTerms, .FlowDiagnostics, .ProgressMessengers
